+++ xolotl/xolotlCore/diffusion/Diffusion1DHandler.h
@@ -99,7 +99,7 @@ public:
 	 * @param iz The position on the z grid
 	 */
 	void computePartialsForDiffusion(IReactionNetwork *network,
-			double *val, int *indices, double hxLeft, double hxRight, int ix,
+			double *val, long int *indices, double hxLeft, double hxRight, int ix,
 			double sy = 0.0, int iy = 0, double sz = 0.0, int iz = 0);
 
 };
